pyproject.toml
[tool.poetry]
name = "eqbackend"
version = "0.1.0"
description = ""
authors = ["MurenMurenus <maximkosmos5360@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
phonenumbers = "^8.13.18"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^7.1.0"}
djoser = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
